refactor: restucture folders for article page
